+++ logging_manager/logging_manager_flutter/lib/logging_manager_flutter.dart
@@ -1,15 +1,13 @@
-import 'dart:async' show Zone, runZonedGuarded;
+import 'dart:async' show Zone;
 import 'dart:isolate';
 
 import 'package:flutter/widgets.dart'
-    show FlutterError, FlutterErrorDetails, WidgetsFlutterBinding;
+    show FlutterErrorDetails, WidgetsFlutterBinding;
 import 'package:logging_manager/logging_manager.dart';
 
 export 'package:logging_manager/logging_manager.dart';
 
-typedef FutureCallback = Future<void> Function();
-
-extension FlutterLoggingManager on LoggingManager {
+extension FlutterLoggingManagerX on LoggingManager {
   /// Creates a log from [FlutterErrorDetails].
   void onFlutterError(FlutterErrorDetails details) {
     final message = details.exceptionAsString();
@@ -24,39 +22,15 @@ extension FlutterLoggingManager on LoggingManager {
     ));
   }
 
-  /// Sends a log to the current logger managed by this manager.
-  Future<void> onRecordError(
-    Object? error,
-    StackTrace? stackTrace, {
-    String? reason,
-    bool fatal = false,
-  }) {
-    if (fatal) {
-      logger.severe(reason, error, stackTrace);
-    } else {
-      logger.warning(reason);
-    }
-
-    /// To support being await-ed when overriding classes does something asynchronously.
-    return Future.value(null);
+  void listenErrorsWithCurrentIsolate() {
+    return listenErrorsInIsolate(Isolate.current);
   }
 
-  void listenErrorsInCurrentIsolate() {
-    Isolate.current.addErrorListener(RawReceivePort((pair) async {
-      final List<dynamic> errorAndStacktrace = pair;
-      await onRecordError(
-        errorAndStacktrace.first,
-        errorAndStacktrace.last,
-        fatal: true,
-      );
-    }).sendPort);
-  }
-
-  Future<void>? runZoneGuardedWithLogging(
+  Future<void>? runFlutterInZoneGuardedWithLogging(
+    FutureCallback onRun, {
     FutureCallback? beforeRun,
-    FutureCallback onRun,
-  ) {
-    return runZonedGuarded<Future<void>>(
+  }) {
+    return runZoneGuardedWithLogging(
       () async {
         /// We must call WidgetsFlutterBinding.ensureInitialized() inside
         /// runZonedGuarded. Error handling wouldn’t work if
@@ -64,16 +38,8 @@ extension FlutterLoggingManager on LoggingManager {
         WidgetsFlutterBinding.ensureInitialized();
 
         if (beforeRun != null) await beforeRun();
-
-        FlutterError.onError = onFlutterError;
-
-        return onRun();
       },
-      (error, stack) => onRecordError(
-        error,
-        stack,
-        fatal: true,
-      ),
+      onRun,
     );
   }
 }