+++ logging_manager/logging_manager_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* Change `FlutterLoggingManager` to an extension of name `FlutterLoggingManagerX`.
+* Add methods `onFlutterError`, `listenErrorsWithCurrentIsolate`, and `runFlutterInZoneGuardedWithLogging`.
+
 ## 1.0.0
 
 * TODO: Describe initial release.